スクールトップのカードを統一
