+++ client/app/schools/[id]/components/courses/CourseCard.tsx
