+++ client/app/schools/[id]/components/courses/PickupCourses.tsx
@@ -1,10 +1,10 @@
-import PickupCourseCard from "./PickupCourseCard";
-import { CourseDetail, SchoolWithCourses } from "@/types/school";
+import CourseCard from "../courses/CourseCard";
+import { CourseListData, SchoolWithCourses } from "@/types/school";
 import Link from "next/link";
 
 interface PickupCoursesProps {
   school: SchoolWithCourses & {
-    pickupCourses: CourseDetail[];
+    pickupCourses: CourseListData[];
   };
 }
 
@@ -16,19 +16,15 @@ const PickupCourses = ({ school }: PickupCoursesProps) => {
     <section className="my-5 max-w-5xl w-full mx-auto px-4">
       <h2 className="text-xl font-bold mb-4 text-gray-700">おすすめコース</h2>
 
-      <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+      <div className="space-y-4">
         {school.pickupCourses.map((course) => (
-          <PickupCourseCard
-            key={course.id}
-            course={course}
-            schoolId={school.id}
-          />
+          <CourseCard key={course.id} course={course} />
         ))}
       </div>
       <Link href={`/schools/${school.id}/courses`} className="block">
         <button className="my-5 max-w-5xl text-sm w-full px-3 py-3 bg-gray-100 text-gray-700 rounded-sm hover:bg-gray-200">
           コース一覧を見る
-        </button>{" "}
+        </button>
       </Link>
     </section>
   );
