+++ client/app/schools/[id]/components/reviews/PickupReviews.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import PickupReviewCard from "./PickupReviewCard";
+import ReviewCard from "./ReviewCard";
 import { ReviewWithUser } from "@/types/review";
 
 export default function PickupReviews({
@@ -25,7 +25,7 @@ export default function PickupReviews({
         <Link href={`/schools/${schoolId}/reviews`} className="block">
           <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
             {reviews.map((review) => (
-              <PickupReviewCard key={review.id} review={review} />
+              <ReviewCard key={review.id} review={review} />
             ))}
           </div>
           <button className="my-5 max-w-5xl text-sm w-full px-3 py-3 bg-gray-100 text-gray-700 rounded-sm hover:bg-gray-200">
